wdio-logger: allow to run package in web context
